Use unsigned int for format paramater
Although the size of an enum is compiler-dependent, it is more likely to
be an unsigned int than a uint8_t.

+++ src/pinblock.c
@@ -600,7 +600,7 @@ int pinblock_decode(
 	size_t pinblock_len,
 	const uint8_t* other,
 	size_t other_len,
-	uint8_t* format,
+	unsigned int* format,
 	uint8_t* pin,
 	size_t* pin_len
 )
